Reintroduce the sponsors list on the bottom of the pages
Closes cloud-native-austria/cloud-native-days-austria#48
