+++ src/components/pages/home/sponsors-list/index.js
@@ -1,0 +1,3 @@
+import SponsorsList from './sponsors-list';
+
+export default SponsorsList;
