+++ src/components/pages/home/sponsors-list/logos/AwsCommunityDays.png
