+++ src/components/pages/home/sponsors-list/logos/CloudWerkstatt.png
