+++ src/components/pages/home/sponsors-list/logos/Cloudbees.png
