+++ src/components/pages/home/sponsors-list/logos/Cloudpilots.png
