+++ src/components/pages/home/sponsors-list/logos/Cncf.png
