+++ src/components/pages/home/sponsors-list/logos/Dynatrace.png
