+++ src/components/pages/home/sponsors-list/logos/Exoscale.png
