+++ src/components/pages/home/sponsors-list/logos/Suse.png
