+++ src/components/pages/home/sponsors-list/logos/Veeam.png
